reorganize some stuff
