+++ src/main/java/com/ghostipedia/cosmiccore/CosmicCore.java
@@ -15,6 +15,7 @@ import com.gregtechceu.gtceu.api.data.chemical.material.event.MaterialRegistryEv
 import com.gregtechceu.gtceu.api.data.chemical.material.registry.MaterialRegistry;
 import com.gregtechceu.gtceu.api.machine.MachineDefinition;
 import com.gregtechceu.gtceu.api.recipe.GTRecipeType;
+import com.gregtechceu.gtceu.api.sound.SoundEntry;
 import com.gregtechceu.gtceu.common.block.CoilBlock;
 import com.gregtechceu.gtceu.config.ConfigHolder;
 import com.lowdragmc.lowdraglib.Platform;
@@ -43,6 +44,7 @@ public class CosmicCore {
        // bus.addGenericListener(Class.class, this::registerRecipeConditions);
        // bus.addGenericListener(MachineDefinition.class, this::registerMachines);
         bus.addGenericListener(MachineDefinition.class, this::registerMachines);
+        bus.addGenericListener(SoundEntry.class, this::registerSounds);
 
         if (Platform.isClient()) {
             bus.register(CosmicCoreClient.class);
@@ -92,6 +94,10 @@ public class CosmicCore {
     public void registerMachines(GTCEuAPI.RegisterEvent<ResourceLocation, MachineDefinition> event) {
         CosmicMachines.init();
         StellarIris.init();
+    }
+
+    public void registerSounds(GTCEuAPI.RegisterEvent<ResourceLocation, SoundEntry> event) {
+        CosmicSounds.init();
     }
 
     @SubscribeEvent
