--- src/main/java/com/ghostipedia/cosmiccore/api/item/component/FluidStats.java
@@ -1,9 +1,0 @@
-package com.ghostipedia.cosmiccore.api.item.component;
-
-import com.gregtechceu.gtceu.api.capability.GTCapabilityHelper;
-
-public class FluidStats {
-    public FluidStats() {
-
-    }
-}
