+++ src/main/java/com/ghostipedia/cosmiccore/client/ForgeClientEventHandler.java
@@ -3,16 +3,11 @@ package com.ghostipedia.cosmiccore.client;
 import com.ghostipedia.cosmiccore.CosmicCore;
 import com.ghostipedia.cosmiccore.client.renderer.StructureBoundingBox;
 import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.api.distmarker.OnlyIn;
 import net.minecraftforge.client.event.RenderLevelStageEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 
-@Mod.EventBusSubscriber(
-        modid = CosmicCore.MOD_ID,
-        bus = Mod.EventBusSubscriber.Bus.FORGE,
-        value = Dist.CLIENT)
-@OnlyIn(Dist.CLIENT)
+@Mod.EventBusSubscriber(modid = CosmicCore.MOD_ID, bus = Mod.EventBusSubscriber.Bus.FORGE, value = Dist.CLIENT)
 public class ForgeClientEventHandler {
 
     @SubscribeEvent
@@ -20,7 +15,6 @@ public class ForgeClientEventHandler {
         var stage = event.getStage();
         if (stage == RenderLevelStageEvent.Stage.AFTER_TRIPWIRE_BLOCKS) {
             StructureBoundingBox.renderStructureSelect(event.getPoseStack(), event.getCamera());
-
         }
     }
 }
