+++ src/main/java/com/ghostipedia/cosmiccore/common/data/CosmicSounds.java
@@ -1,14 +1,13 @@
-//package com.ghostipedia.cosmiccore.common.data;
-//import com.gregtechceu.gtceu.api.registry.GTRegistries;
-//import com.gregtechceu.gtceu.api.sound.SoundEntry;
-//
-//import static com.ghostipedia.cosmiccore.api.registries.CosmicRegistration.REGISTRATE;
-//
-//
-//public class CosmicSounds {
-//    static {
-//        GTRegistries.SOUNDS.unfreeze();
-//    }
-//    public static final SoundEntry BLACK_HOLE_CRY = REGISTRATE.sound("black_hole_cry").build();
-//    public static void init() {}
-//}
+package com.ghostipedia.cosmiccore.common.data;
+
+import com.ghostipedia.cosmiccore.CosmicCore;
+import com.gregtechceu.gtceu.api.registry.registrate.SoundEntryBuilder;
+import com.gregtechceu.gtceu.api.sound.SoundEntry;
+
+public class CosmicSounds {
+
+    public static final SoundEntry BLACK_HOLE_CRY = new SoundEntryBuilder(CosmicCore.id("black_hole_cry")).build();
+
+    public static void init() {
+    }
+}
