+++ src/main/java/com/ghostipedia/cosmiccore/forge/ABSModifications.java
@@ -4,7 +4,6 @@ import com.ghostipedia.cosmiccore.CosmicCore;
 import com.ghostipedia.cosmiccore.common.data.materials.CosmicMaterials;
 import com.gregtechceu.gtceu.api.data.chemical.material.event.PostMaterialEvent;
 import com.gregtechceu.gtceu.api.data.chemical.material.properties.PropertyKey;
-import com.gregtechceu.gtceu.common.data.GTMaterials;
 import com.gregtechceu.gtceu.data.recipe.misc.alloyblast.CustomAlloyBlastRecipeProducer;
 import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -12,6 +11,7 @@ import net.minecraftforge.fml.common.Mod;
 
 @Mod.EventBusSubscriber(modid = CosmicCore.MOD_ID, bus = Mod.EventBusSubscriber.Bus.MOD)
 public class ABSModifications {
+
     @SubscribeEvent(priority = EventPriority.LOW)
     public static void addAlloyBlastProperties(PostMaterialEvent event) {
         CosmicMaterials.ResonantVirtueMeld.getProperty(PropertyKey.ALLOY_BLAST)
